+++ src/api/types/wa.types.ts
@@ -146,4 +146,5 @@ export const MessageSubtype = [
 export const Integration = {
   WHATSAPP_BUSINESS: 'WHATSAPP-BUSINESS',
   WHATSAPP_BAILEYS: 'WHATSAPP-BAILEYS',
+  EVOLUTION: 'EVOLUTION',
 };